feat: Ability to change output format with "in <format>"
This introduces a new operator called "in", which currently is only used
to convert between number formats. This enables the user to change the
output number's format by appending "in <decimal/hex/binary>" to the end
of the line. "decimal" is the default.

NOTE: This operator is allowed anywhere in the line, however it only
actually takes affect at the end of the line. In the future, we might
want to deduce the output format from the format used in the literals in
the line (e.g. 80% hex literals -> hex output).