feat: Add Groups
This adds groups to the calculator. Groups are parts of the calculation
that are surrounded by brackets ("()"), which are evaluated before
anything else. They can be nested, and every opening bracket must have a
corresponding closing bracket.

Currently, they act just the same as literals, but in the future we will
support inferred multiplication (e.g. "2(3 + 3)" = 2 * (3 + 3) => 12)