feat: Add support for simple units
This adds support for simple units in the calculator. If a number is
suffixed with an identifier, the parser checks whether it is a valid
unit, and if so, stores it in the AstNode. In calculations, when two
different units are used, rhs' unit is converted into lhs' unit
beforehand.

NOTE: For now, only conversion during calculations is supported. In the
future, the "in" operator will be used to force conversion.