feat(GUI): Add syntax highlighting in the input field
The equations typed in the input field in the GUI now receive syntax
highlighting information from the calculator.

For this to get working, the tokenizer had to be changed again to work
with bytes instead of chars, since the Token's ranges should start and
end at the actual char boundaries of the string instead of being indices
into the string's char array. Therefore, for non-ascii characters, their
individual bytes have to be inserted into the string of allowed
characters, to still match with the bytes in String::contains().
These are not the unicode codes (e.g. "°"'s unicode is U+00B0, but it's
bytes are [0xC2, 0xB0]).