fix(GUI): Handle errors when indices are not at char boundaries
Because of the way the text is rendered in the GUI application, the
layouter is called before we re-calculate (and update syntax
highlighting data) the input. Therefore, the syntax highlighting data
might be wrong in the layouter, causing panics. The layouter has to try
to gracefully handle these errors, since they will be corrected in the
same frame later down the file.
The layouter will now break when it encounters indices that are not at
char boundaries, since using them in the output ranges would yield
panics.