refactor: Move variables and functions into Environment
There is now a new struct called Environment, which stores all
information regarding variables and functions, along some functions for
validating them and resolving them to a value. Additionally, the user
now has to store said environment somewhere and pass it to the calculate
function.
This is a pre-step in allowing custom variables and functions, since
they will then be stored in the environment.