feat: Add custom variables
The calculator now supports declaring custom variables with the syntax
"name := expr". If such a line is encountered, the parser extracts the
name and parses the expression, and returns it as a VariableDeclaration
result. The ast parsed from the expresion is then evaluated, and the
result + unit along with the name from the LHS is stored in the
Environment.
If RHS is empty, the variable is removed.