feat: Custom functions
Using the same ":=" token as for defining variables, one can now define
custom functions. The syntax is as follows:

name(arg1, arg2, ...) := expr

In "expr" the names of the arguments given inside the brackets can be
used. The function can then be invoked just like any other function. If
"expr" is empty, the function is removed.