feat: Add basic support for currencies
On build, the file src/calculator/environment/default_currencies.rs is
now generated. It contains default conversion numbers for currencies
that are used to convert between currencies. Since these conversion
numbers are based on a base currency ("EUR"), the conversion works as
follows:
1. Convert the number into the base currency by dividing by the number
   stored for the source currency.
2. Convert the number into the destination currency by multiplying with
   the number stored for the destination currency.

In the future, this conversion data should be downloaded from the API
as soon as the app starts to have the most up-to-date data. These
default conversions are then going to be used as a fallback.