feat(currencies): Save currencies in local file
Everytime the calculator starts, it now tries to load currency
conversion data from a local file. If this fails / the file does not
exist, it falls back to the default data in the convert() function.
After that, we try to asyncronously download new currency data from the
API, update the struct fields and save them to a file.