feat: Currency support
This adds support for currencies as units. The application now fetches 
the currency conversions from an API (exchangerate.host) on startup and 
stores them on desktop (see FIXMEs for WASM). Default conversions are 
also generated in build.rs during build time, stored in the file 
environment/default_currencies.rs and act as a fallback.