feat(GUI): Allow plotting re-definitions of functions
When a function is re-defined, you can now still plot older versions of
the function.
For this to work, the function field in the Line::Line variant now stores
the Function struct from the environment, which has been made public
(although not the fields). When plotting, we now just ask the environment
to resolve that function struct.