feat: Feature to solve linear equations
The calculator can now solve linear equations if there is a question
mark on either side of the equals sign.

Question mark position restrictions:
- It or its enclosing groups cannot be surrounded by a power sign, since
  that would make the equation non-linear
- It cannot be a function argument

The process is noted down in the source code, however the general idea
is that we're re-arranging the equation to equal 0, since then the
x-coordinate of the intersection point with the x-axis is the value
we're looking for. Then, we just calculate the values m, c from y = mx +
c and find the value of x where y = 0.