fix(GUI): Don't crash if function is not defined for x-value
If we get an error trying to resolve the function in the plot drawing
function, it is likely caused by the function not being defined for this
x-value (since syntax errors are catched during parsing). To prevent
crashing on the .unwrap(), we now just return infinity if we get an
error (see comment).

Example:
f(x) := 1 / x ^ 2    => not defined for x = 0 (DivideByZero)