feat(discord): Add sessions
The user can now create a "session" using the command
`/session new <name>`. This will create a new thread with the given name.
The calculator will use the same environment for any `/calc` invocations in
the thread, allowing you to for example define functions and use them
later.
Sessions can be deleted with the command `/session delete <name>`.
<name> can be omitted if the command is invoked from inside the session
thread, in which case the current session is deleted.