feat(discord): Persist sessions across bot restarts
The sessions for each guild are now saved to a file (the name being
the id of the guild) using the `serde` and `ron` crates. On startup, the
bot tries to load the sessions from the files, to prevent having
meaningless threads in channels when the bot crashes / internet cuts out
etc.