fix: Fix crash in trunk because of gui's build.rs
When using trunk on windows, `cfg(windows)` is true in the gui's
build.rs, meaning we were trying to compile windows assets into the
binary, leading to trunk crashing. We now make sure that we're not on
wasm by checking if the `TARGET` environment variable contains "wasm32",
and only if not, we compile in the resources.