refactor(GUI): Rewrite output_text() function and get rid of sliding
The output_text() function has now been rewritten, to simplify it and
thereby getting rid of the text sliding. It has never worked properly
before due to what I think is that the ui didn't update often enough
for the text to slide properly. Now, we just right-align the text and
clip it to what we have available.