fix(tokenizer): Remove sign detection
This is already implemented in the parser as modifiers, since we need it
for things that are not literals (e.g. group / function calls), so we
don't need it in the tokenizer.
Additionally, this also enables e.g. `1+2` without spaces, which was not
possible before due to the tokenizer thinking the plus was a sign.