feat: Allow formatting output numbers with '_' as the thousand separator
By giving the -s argument to the CLI, or enabling it through the
settings window in the GUI, the user can now choose whether the output
of calculations should be formatted with a thousands separator ('_') or
not. For decimal numbers, this will be done in packets of 3, and for
binary and hexadecimal numbers in packets of 4.