feat(gui): input formatting
The text in the input field can now be formatted using the Cmd+alt+L
shortcut. We do this by calling the format() function on the calculator
for every line and collecting the results, eventually replacing what was
in source previously.
To format, we tokenize the input and add spaces depending on which
token we encounter and a few additional rules.