feat(GUI): add line picker dialog
Using the button in the "Navigate" menu, or the shortcut Cmd+G, the user
can now bring up a text edit dialog, in which they can type a line
number to put their cursor at the start of. While this dialog is open,
the rest of the UI is disabled. It can either be cancelled using escape,
or confirmed with enter. If the string typed in is not a number, the
text is displayed red and pressing enter does nothing.