fix(GUI): allow choosing which redefinitions should be shown in the plot
The user can now choose whether the function should be shown in the plot
per redefinition. To allow that, we store the function name, as well as
whether it should be shown in the functions Vec in update_lines. Then,
we search for the first occurrence of the function name and take the
associated show-value. Then we remove it. The next time we search for
that function, we will get the next entry.