feat(GUI): add fullscreen plot
In the top right hand corner of the plot, there is now a button which
opens the same plot in fullscreen on top of everything else. This is
done by making an egui::Window, with no border and title, and sizing it
to the window's size. While the fullscreen plot is open, the underlying
plot is not drawn to save resources.