fix(GUI): make line picker + format shortcuts work when input is not focused
There is now a distinction between a text edit shortcut and a normal
shortcut. Text edit shortcuts need the current cursor range, meaning
the text edit has to be at least focused for them to work. Other
shortcuts, like the line picker shortcut, don't need the text edit to be
focused, which this commit now allows.
In additon, the line picker dialog call has been moved further up in the
App::update() function to make it centered inside the screen.