fix(GUI): make the input layouter correct temporary errors in color segments
When the user is typing into the input text field, the input layouter is
called before update_lines, making the color segments temporarily
invalid. This causes issues when there are non-ascii characters in the
input string, as a range might falsly end within the non-ascii
character. The input layouter now corrects for this by incrementing the
start and end indices of every LayoutSection until they are both at char
bounds of the full string, if necessary.