fix(GUI): also show "Copied!" tooltip when line does not fill overflow
When an output line overflows the width of the output section, we show a
tooltip on hover so that the user can see the entire output, as well as
a second line saying "Copied!" if the user clicked on the text. Now,
this "Copied!" text also shows if the output does not overflow.