refactor(parser): new parser architecture
In this commit, the parser has been rewritten to have a similar
structure to the tokenizer. Previously, we would look at the line token
by token, however now, we look at a line in "blocks". A block might be a
sequence of modifiers, a group, a function call, etc.

The rules defining which of these blocks can come in which order is embedded
in the code by calling the corresponding "accept_*" functions after one
another. To to this, it is important to realise that a line always has
the same structure, i.e. <number> (<operator> <number>)* ('*' means 0+
times). Here, <number> does not mean literal, but rather anything that
can be resolved to a literal during evaluation (e.g. groups are also
"numbers" (kind of a bad word though)).

This gives us a lot more freedom about what we can parse (see maybe
"objects" for dates, ...) and makes it clearer what is allowed when,
since these rules aren't defined in a single function and then assumed
everywhere else.