chore(calculator): remove "strum" and "rust_decimal" dependencies
"strum" was only needed because of its EnumIter in the old parser,
however the new parser does not need this anymore.

"rust_decimal" was only used to round a number to a certain number of
decimal places, however there is a better way to do this (see
implementation of builtin "round(n, dp)" function), which is now also
used to round the result of calculations.
This interestingly also fixed an issue where we could have -0.0000000000
as the result.