refactor(parser): don't use sub-parsers for groups
The parser now doesn't create new parsers to parse the contents of
groups. Instead, we save a "stack" of Vec<AstNode>, which are the
AstNodes we have parsed so far from a group. It then appends new parsed
nodes to the last vector in that stack. When the parser now encounters
an open bracket, it simply pushes an empty vector to said stack, thereby
increasing the nesting level by one. When a close bracket is
encountered, the last vector is popped off of the stack and put into a
group AstNode, before being pushed to the new last vector, thereby
decreasing the nesting level by one.

This makes it easier to save a "parsing state", making it later possible
to continue parsing on a new line.