feat: basic object system
This introduces a new concept into the calculator called "objects".
Objects are created with the syntax `{<name> <args>}`, where <args>
could be anything and it is up to the object to parse it. The objects
then define operations that can be done with them and calculate them.
This allows for more complicated data structures.

A date object has been started in this commit as well, however it is
very basic and needs refining.