chore(colors): redo color generation
Previously, we would randomly generate colors for bracket pairs and try
to restore the rng's seed to get the same colors every time. This did
not work flawlessly however and resulted in the colors not being
consistent across different devices. Now, we have a pre-defined pallete
of 5 colors, which are used for bracket pairs. These are also lightened
by a multiple of 20% the greater the nesting level gets. This makes the
colors consistent across all devices as well as avoids accidentally
generating colors that are too dark to see.