feat(DateObject): allow using calculations in the arguments
Instead of having to write the date out as a string, the user can now
use [...] to insert a calculation where a number would have been. The
dots between the [] / numbers are still mandatory. Between the square
brackets, the user can type anything they are also able to type outside
of the object, as long as it evaluates to a number. The number is then
used in that place.
For this to work, the tokenizer does some very naive looking at the
tokens and figures out whether it's in an object. Then, it produces
ObjectArgs and Open-/CloseBracket tokens, and normal ones between the
brackets. The parser then collects the tokens between the brackets into
a sub-AST, as well as extracting the string from the ObjectArgs tokens
and passes this information to the DateObject for it to parse it. The
DateObject immediately evaluates the ASTs if there are any, which means
we have to pass an environment and currency information. Evaluating
these on the spot reduces more complexity since we would have to
deal with individual fields that are either ASTs or numbers (e.g. it is
impossible to implement PartialOrd (/ Ord) for an AST).