fix(parser): wrong error when using a definition sign in brackets
Previously, when trying to parse a function definition head (the part
before the definition sign), we would first check if there is a
definition sign _anywhere_, and then start emitting errors when we don't
find the right tokens. This lead to the wrong errors being thrown when
putting the definition sign a nesting level deeper (e.g. `(v := 3)`).
Instead, the try_accept_function_definition_head function now parses the
entire definition head, recovering from errors and storing the first
error it encountered. If it does not find a definition head, it
discards the error, resets self.index and returns None, thereby
"passing" parsing on to the accept_expression function.
If it finds a definition sign however, it returns the first error stored
earlier, if there is one, and the parsed data otherwise.
This way we know whether the definition sign is deeper nested and can
emit the correct errors.