refactor: Context struct
This adds a struct called Context, which holds references to the
environment, Currencies and Settings. This makes it easier to pass these
down to other components of the calculator, since we now only need one
function argument. In addition, this struct can be copied, meaning we
don't even have to borrow this struct.