feat(cli): allow changing settings using "set"
When typing "set" into the input of the cli, the application now expects
a "path" to a setting, along with the new value. The "path" follows the
dot notation to get to the setting, as if it was typed in Rust, relative
to the Settings struct (e.g. "calculator.settings.date.format" is "date.format").

The syntax is as follows:
set (item).+ = value

The CLI also persists the settings using the RON crate.