revert: "experiment(GUI): value table"
This reverts commit ec211db956e30569fc217fc829ef02bd7bd0125a.

The implementation of this feature had too many issues correctly sizing
the table, because ui.available_height() DOES NOT RETURN THE CORRECT
HEIGHT >:C. This makes it impossible to correctly determine the
necessary amount of rows to draw, meaning the bottom panel jumps
around because after it snaps to the height we occupy, there (according
to ui.available_height()) is space for another row. After we add that
row the bottom panel jumps again and this repeats forever, in both
directions.