fix(gui): incorrect line numbers when lines are wrapping
Previously, there was a bug where if a line wrapped two times, the next
line number would be on the third line, but it should be on the fourth
because the first line is taking up three lines. This was caused by us
not adding enough newline characters to App::line_numbers_text in
App::update_lines. In addition, the output line numbers are now always
at the first line of a wrapping line.