feat(gui): align line numbers to the right
The line numbers are now aligned to the right, putting them closer to
the input text field. This is achived by padding the lines with spaces
if they're not long enough. This is not optimal, however in the current
layout, it is difficult to let egui right-align the text, so we use this
workaround.