fix(units): allow Unit::simplify() to remove itself
Unit::simplify() now returns a boolean, indicating whether the unit
should be kept after calling the function. An example is the unit `h/h`,
which is equal to one. In such a case, the function will return false,
indicating that the unit should not exist afterwards anymore. In the
parser, we just return none from the corresponding accept functions and
in the AST, we remove it from the AstNode.