fix(parser): prefer units over variables
When typing a division operator after a unit and the variable on the
RHS is also a unit, the parser previously would make the whole triplet
a fractional unit. Instead, we now prefer variables over units, so
if the denominator is a variable, we don't parse it into a unit.