feat(calculator): postfix assignment
Using the syntax `<expr> =: <variable name/function definition head>`,
one can now create variables / functions too. This might be useful when
having typed a long calculation and only at the end realising that the
result needs to be in a variable. However, it is sort of an experiment
and might get removed again in the future if it is not useful.