refactor: use custom range struct
In this refactor, the codebase has been changed to use a custom range
struct called SourceRange. This is necessary, since if we want to
introduce multiline parsing, we need to be aware of which line we're on.
Therefore, this new SourceRange struct already has fields for a start
and end line, which will be used in the future.