refactor(parser): simplify parse_single() function
The accept_expression() function now takes a boolean for whether to
accept a definition start, i.e. a number before the actual while loop.
Therefore, this does not need to be done in the parse_single() function,
since there, this needs to be done in three places (after opening
brackets, after function variant heads and before the while loop.