fix(parser): don't pass boolean operator to sub parser
Previously, because of a refactor in the parser, we would pass the
boolean operator to sub parsers, causing them to return
ParserResultData::BooleanExpression, which was not desired. Instead,
each parser now manages its own boolean operator. This still works,
since the nesting level will automatically increase for sub parsers and
boolean operators are currently only allowed at the top level.