fix(frontend): correctly react to changes from the custom keyboard
When the text in the TextField changed only via the
TextEditingController, the TextField doesn't call its onChanged method.
Now, we directly register ourselves as a listener on the
TextEditingController for the input TextField to receive all updates. In
addition, we also don't store the input as a separate variable anymore,
since it can easily be retrieved via the TextEditingController.