fix(frontend): clear color segments when there's an error
When the calculator returns an error, there might still be normal color
segments (specifically when the Engine returned and error and not the
Parser/Tokenizer). The app now ignores those color segments though,
since when there's an error we want to keep the line bland so that it is
more obvious that there was an error.