fix(frontend): use correct plot size in PlotWidget
Previously, PlotWidget used the screen sizes for its calculations.
However, since the plot is not the same size as the screen, this caused
e.g. the plot to move slower vertically or the y-Axis to not have the
same tick distances as the x-Axis. Now, we use a LayoutBuilder around
the CustomPaint and GestureDetector widgets and use the maxWidth and
maxHeight fields on the constraints it gives us to do the calculation
instead.