feat(frontend): add API for drawing graphs in the plot
The plot now accepts a list of PlotGraph objects, which it will draw
into the coordinate system. Each PlotGraph object contains a function
taking the x-coordinate and returning the y-coordinate for that point.
The plot painter then calls that function repeatedly, before drawing
lines to connect the points in the plot.