feat(frontend): draw calculator functions in the plot
When declaring a function, the last definition of that function can now
be seen in the plot. There, visibility can be toggled using the legend.

NOTE: We now copy the SourceRange into a dart class when creating a
StyleSegment, since we are not deallocating the results correctly, which
caused the ranges to contain weird values.