feat(frontend): show plot on HomePage if there is enough space
If there is enough space on the HomePage, the plot is now shown to the
right of the results panel. For this to work, the plot now recalculates
its y-Axis automatically whenever the available space changes, so that
one unit on the x-Axis equals the same amount of pixels on the y-Axis.