feat(frontend): add MenuBar with editing actions
The HomePage now contains a MenuBar above the input text field, with
which the editing actions format, surround selection with brackets and
copy result are implemented. This way, it can be used just like in other
applications (and also like in the egui version), as well as having
shortcuts working.